+++ include/UDRefl/Object.h
@@ -154,6 +154,9 @@ namespace Ubpa::UDRefl {
 		// self [r] vars and all bases' [r] vars
 		void ForEachRVar(const std::function<bool(TypeRef, FieldRef, ConstObjectPtr)>& func) const;
 
+		// self [r] owned vars and all bases' [r] owned vars
+		void ForEachROwnedVar(const std::function<bool(TypeRef, FieldRef, ConstObjectPtr)>& func) const;
+
 		//
 		// Algorithm
 		//////////////
@@ -166,12 +169,15 @@ namespace Ubpa::UDRefl {
 		std::vector<MethodRef>                                     GetMethods();
 		std::vector<std::tuple<TypeRef, FieldRef, ConstObjectPtr>> GetTypeFieldRVars();
 		std::vector<ConstObjectPtr>                                GetRVars();
+		std::vector<std::tuple<TypeRef, FieldRef, ConstObjectPtr>> GetTypeFieldROwnedVars();
+		std::vector<ConstObjectPtr>                                GetROwnedVars();
 
-		std::optional<TypeID   > FindTypeID(const std::function<bool(TypeID        )>& func) const;
-		std::optional<TypeRef  > FindType  (const std::function<bool(TypeRef       )>& func) const;
-		std::optional<FieldRef > FindField (const std::function<bool(FieldRef      )>& func) const;
-		std::optional<MethodRef> FindMethod(const std::function<bool(MethodRef     )>& func) const;
-		ConstObjectPtr           FindRVar  (const std::function<bool(ConstObjectPtr)>& func) const;
+		std::optional<TypeID   > FindTypeID   (const std::function<bool(TypeID        )>& func) const;
+		std::optional<TypeRef  > FindType     (const std::function<bool(TypeRef       )>& func) const;
+		std::optional<FieldRef > FindField    (const std::function<bool(FieldRef      )>& func) const;
+		std::optional<MethodRef> FindMethod   (const std::function<bool(MethodRef     )>& func) const;
+		ConstObjectPtr           FindRVar     (const std::function<bool(ConstObjectPtr)>& func) const;
+		ConstObjectPtr           FindROwnedVar(const std::function<bool(ConstObjectPtr)>& func) const;
 
 		DereferenceProperty GetDereferenceProperty() const;
 		TypeID              DereferenceID() const;
@@ -432,10 +438,17 @@ namespace Ubpa::UDRefl {
 		// self [r/w] vars and all bases' [r/w] vars
 		void ForEachRWVar(const std::function<bool(TypeRef, FieldRef, ObjectPtr)>& func) const;
 
+		// self [r/w] owned vars and all bases' [r/w] owned vars
+		// if obj is &{const{T}}, then return directly
+		void ForEachRWOwnedVar(const std::function<bool(TypeRef, FieldRef, ObjectPtr)>& func) const;
+
 		std::vector<std::tuple<TypeRef, FieldRef, ObjectPtr>> GetTypeFieldRWVars();
 		std::vector<ObjectPtr>                                GetRWVars();
+		std::vector<std::tuple<TypeRef, FieldRef, ObjectPtr>> GetTypeFieldRWOwnedVars();
+		std::vector<ObjectPtr>                                GetRWOwnedVars();
 
-		ObjectPtr FindRWVar(const std::function<bool(ObjectPtr)>& func) const;
+		ObjectPtr FindRWVar     (const std::function<bool(ObjectPtr)>& func) const;
+		ObjectPtr FindRWOwnedVar(const std::function<bool(ObjectPtr)>& func) const;
 
 		//
 		// Meta